% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{isProb}
\alias{isProb}
\title{Check if a Valid Probability}
\usage{
isProb(prob)
}
\arguments{
\item{prob}{the probability to check}
}
\value{
TRUE if conditions are met, FALSE if otherwise
}
\description{
Checks if the input vector has a single entry that is between 0 and 1
}
