% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{isValidDataVector}
\alias{isValidDataVector}
\title{Is Vector Valid Data?}
\usage{
isValidDataVector(x)
}
\arguments{
\item{x}{the vector to be tested}
}
\value{
TRUE or FALSE
}
\description{
Determines if input vector is a valid vector of real valued observations
}
