% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{isDiscrete}
\alias{isDiscrete}
\title{Determine if input data is discrete}
\usage{
isDiscrete(x)
}
\arguments{
\item{x}{a vector which should be determined if discrete or not.}
}
\value{
the best judgement of whether or not the data was discrete
}
\description{
Attempts to determine if the input data is from a discrete distribution. Will
return true if the data type is of type integer or there are non-unique
values.
}
