% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{print.tstest}
\alias{print.tstest}
\title{Print Tau* Test Results}
\usage{
\method{print}{tstest}(x, ...)
}
\arguments{
\item{x}{the tstest object to be printed}

\item{...}{ignored.}
}
\description{
A simple print function for tstest (Tau* test) objects.
}
